+++ src/Dracodeum/Kit/Primitives/Dictionary.php
@@ -582,21 +582,18 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 	 * <p>The value to evaluate (validate and sanitize).</p>
 	 * @param \Dracodeum\Kit\Primitives\Dictionary|null $template [default = null]
 	 * <p>The template instance to clone from and evaluate into.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same pairs 
-	 * and evaluator functions, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same pairs and evaluator functions.</p>
 	 * @param bool $nullable [default = false]
 	 * <p>Allow the given value to evaluate as <code>null</code>.</p>
 	 * @return bool
 	 * <p>Boolean <code>true</code> if the given value was successfully evaluated into an instance.</p>
 	 */
 	final public static function evaluate(
-		&$value, ?Dictionary $template = null, ?bool $clone_recursive = null, bool $nullable = false
+		&$value, ?Dictionary $template = null, bool $clone = false, bool $nullable = false
 	): bool
 	{
-		return self::processCoercion($value, $template, $clone_recursive, $nullable, true);
+		return self::processCoercion($value, $template, $clone, $nullable, true);
 	}
 	
 	/**
@@ -612,11 +609,8 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 	 * <p>The value to coerce (validate and sanitize).</p>
 	 * @param \Dracodeum\Kit\Primitives\Dictionary|null $template [default = null]
 	 * <p>The template instance to clone from and coerce into.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same pairs 
-	 * and evaluator functions, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same pairs and evaluator functions.</p>
 	 * @param bool $nullable [default = false]
 	 * <p>Allow the given value to coerce as <code>null</code>.</p>
 	 * @throws \Dracodeum\Kit\Primitives\Dictionary\Exceptions\CoercionFailed
@@ -625,10 +619,10 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 	 * If nullable, then <code>null</code> may also be returned.</p>
 	 */
 	final public static function coerce(
-		$value, ?Dictionary $template = null, ?bool $clone_recursive = null, bool $nullable = false
+		$value, ?Dictionary $template = null, bool $clone = false, bool $nullable = false
 	): ?Dictionary
 	{
-		self::processCoercion($value, $template, $clone_recursive, $nullable);
+		self::processCoercion($value, $template, $clone, $nullable);
 		return $value;
 	}
 	
@@ -645,11 +639,8 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 	 * <p>The value to process (validate and sanitize).</p>
 	 * @param \Dracodeum\Kit\Primitives\Dictionary|null $template [default = null]
 	 * <p>The template instance to clone from and coerce into.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same pairs 
-	 * and evaluator functions, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same pairs and evaluator functions.</p>
 	 * @param bool $nullable [default = false]
 	 * <p>Allow the given value to coerce as <code>null</code>.</p>
 	 * @param bool $no_throw [default = false]
@@ -659,8 +650,7 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 	 * <p>Boolean <code>true</code> if the given value was successfully coerced into an instance.</p>
 	 */
 	final public static function processCoercion(
-		&$value, ?Dictionary $template = null, ?bool $clone_recursive = null, bool $nullable = false,
-		bool $no_throw = false
+		&$value, ?Dictionary $template = null, bool $clone = false, bool $nullable = false, bool $no_throw = false
 	): bool
 	{
 		//nullable
@@ -683,24 +673,12 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 			//object
 			if (is_object($value) && $value instanceof Dictionary) {
 				if (isset($template)) {
-					//initialize
 					$instance = $template->clone()->clear();
-					$keys = $value->keys;
-					$values = $value->values;
-					
-					//clone
-					if ($clone_recursive === true) {
-						$keys = UType::cloneValue($keys);
-						$values = UType::cloneValue($values);
-					}
-					
-					//set
-					foreach ($keys as $index => $key) {
-						$instance->set($key, $values[$index]);
+					foreach ($value->keys as $index => $key) {
+						$instance->set($key, $value->values[$index]);
 					}
 					$value = $instance;
-					
-				} elseif ($clone_recursive !== null) {
+				} elseif ($clone) {
 					$value = $value->clone();
 				}
 				return true;
@@ -709,9 +687,6 @@ IArrayable, IArrayInstantiable, IKeyable, IStringable, ICloneable
 			//array
 			$array = $value;
 			if (UData::evaluate($array)) {
-				if ($clone_recursive === true) {
-					$array = UType::cloneValue($array);
-				}
 				$value = isset($template) ? $template->clone()->setAll($array) : static::build($array);
 				return true;
 			}