+++ src/Dracodeum/Kit/Structure.php
@@ -285,11 +285,8 @@ IArrayInstantiable, ICloneable
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param mixed $value [reference]
 	 * <p>The value to evaluate (validate and sanitize).</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same 
-	 * properties, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same properties.</p>
 	 * @param callable|null $builder [default = null]
 	 * <p>The function to use to build an instance with a given set of properties.<br>
 	 * It is expected to be compatible with the following signature:<br>
@@ -310,10 +307,10 @@ IArrayInstantiable, ICloneable
 	 * <p>Boolean <code>true</code> if the given value was successfully evaluated into an instance.</p>
 	 */
 	final public static function evaluate(
-		&$value, ?bool $clone_recursive = null, ?callable $builder = null, bool $nullable = false
+		&$value, bool $clone = false, ?callable $builder = null, bool $nullable = false
 	): bool
 	{
-		return self::processCoercion($value, $clone_recursive, $builder, $nullable, true);
+		return self::processCoercion($value, $clone, $builder, $nullable, true);
 	}
 	
 	/**
@@ -327,11 +324,8 @@ IArrayInstantiable, ICloneable
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param mixed $value
 	 * <p>The value to coerce (validate and sanitize).</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same 
-	 * properties, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same properties.</p>
 	 * @param callable|null $builder [default = null]
 	 * <p>The function to use to build an instance with a given set of properties.<br>
 	 * It is expected to be compatible with the following signature:<br>
@@ -354,10 +348,10 @@ IArrayInstantiable, ICloneable
 	 * If nullable, then <code>null</code> may also be returned.</p>
 	 */
 	final public static function coerce(
-		$value, ?bool $clone_recursive = null, ?callable $builder = null, bool $nullable = false
+		$value, bool $clone = false, ?callable $builder = null, bool $nullable = false
 	): ?Structure
 	{
-		self::processCoercion($value, $clone_recursive, $builder, $nullable);
+		self::processCoercion($value, $clone, $builder, $nullable);
 		return $value;
 	}
 	
@@ -372,11 +366,8 @@ IArrayInstantiable, ICloneable
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param mixed $value [reference]
 	 * <p>The value to process (validate and sanitize).</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same 
-	 * properties, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same properties.</p>
 	 * @param callable|null $builder [default = null]
 	 * <p>The function to use to build an instance with a given set of properties.<br>
 	 * It is expected to be compatible with the following signature:<br>
@@ -400,8 +391,7 @@ IArrayInstantiable, ICloneable
 	 * <p>Boolean <code>true</code> if the given value was successfully coerced into an instance.</p>
 	 */
 	final public static function processCoercion(
-		&$value, ?bool $clone_recursive = null, ?callable $builder = null, bool $nullable = false,
-		bool $no_throw = false
+		&$value, bool $clone = false, ?callable $builder = null, bool $nullable = false, bool $no_throw = false
 	): bool
 	{
 		//builder
@@ -428,9 +418,6 @@ IArrayInstantiable, ICloneable
 					$properties = static::getStringProperties($value);
 				} elseif (is_array($value)) {
 					$properties = $value;
-					if ($clone_recursive === true) {
-						$properties = UType::cloneValue($properties);
-					}
 				} elseif (is_callable($value)) {
 					$properties = static::getCallableProperties($value);
 				}
@@ -440,21 +427,13 @@ IArrayInstantiable, ICloneable
 				$instance = $value;
 				if ($instance instanceof Structure) {
 					if (!UType::isA($instance, static::class)) {
-						$properties = $instance->getAll(true);
-						if ($clone_recursive === true) {
-							$properties = UType::cloneValue($properties);
-						}
-						$value = UType::coerceObject($builder($properties), static::class);
-					} elseif ($clone_recursive !== null) {
+						$value = UType::coerceObject($builder($instance->getAll(true)), static::class);
+					} elseif ($clone) {
 						$value = $value->clone();
 					}
 					return true;
 				} elseif (UData::evaluate($instance)) {
-					$properties = $instance;
-					if ($clone_recursive === true) {
-						$properties = UType::cloneValue($properties);
-					}
-					$value = UType::coerceObject($builder($properties), static::class);
+					$value = UType::coerceObject($builder($instance), static::class);
 					return true;
 				}
 			}