+++ src/Dracodeum/Kit/Traits/Evaluators.php
@@ -938,11 +938,8 @@ trait Evaluators
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param string $class
 	 * <p>The class to use.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same 
-	 * properties, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same properties.</p>
 	 * @param callable|null $builder [default = null]
 	 * <p>The function to use to build an instance with a given set of properties.<br>
 	 * It is expected to be compatible with the following signature:<br>
@@ -961,10 +958,10 @@ trait Evaluators
 	 * <p>This instance, for chaining purposes.</p>
 	 */
 	final public function setAsOptions(
-		string $class, ?bool $clone_recursive = null, ?callable $builder = null, bool $nullable = false
+		string $class, bool $clone = false, ?callable $builder = null, bool $nullable = false
 	): object
 	{
-		$this->getEvaluatorsManager()->setAsOptions($class, $clone_recursive, $builder, $nullable);
+		$this->getEvaluatorsManager()->setAsOptions($class, $clone, $builder, $nullable);
 		return $this;
 	}
 	
@@ -979,11 +976,8 @@ trait Evaluators
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param string $class
 	 * <p>The class to use.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same 
-	 * properties, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same properties.</p>
 	 * @param callable|null $builder [default = null]
 	 * <p>The function to use to build an instance with a given set of properties.<br>
 	 * It is expected to be compatible with the following signature:<br>
@@ -1004,10 +998,10 @@ trait Evaluators
 	 * <p>This instance, for chaining purposes.</p>
 	 */
 	final public function setAsStructure(
-		string $class, ?bool $clone_recursive = null, ?callable $builder = null, bool $nullable = false
+		string $class, bool $clone = false, ?callable $builder = null, bool $nullable = false
 	): object
 	{
-		$this->getEvaluatorsManager()->setAsStructure($class, $clone_recursive, $builder, $nullable);
+		$this->getEvaluatorsManager()->setAsStructure($class, $clone, $builder, $nullable);
 		return $this;
 	}
 	
@@ -1022,21 +1016,18 @@ trait Evaluators
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param \Dracodeum\Kit\Primitives\Dictionary|null $template [default = null]
 	 * <p>The template instance to clone from and evaluate into.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same pairs 
-	 * and evaluator functions, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same pairs and evaluator functions.</p>
 	 * @param bool $nullable [default = false]
 	 * <p>Allow a value to evaluate as <code>null</code>.</p>
 	 * @return $this
 	 * <p>This instance, for chaining purposes.</p>
 	 */
 	final public function setAsDictionary(
-		?Dictionary $template = null, ?bool $clone_recursive = null, bool $nullable = false
+		?Dictionary $template = null, bool $clone = false, bool $nullable = false
 	): object
 	{
-		$this->getEvaluatorsManager()->setAsDictionary($template, $clone_recursive, $nullable);
+		$this->getEvaluatorsManager()->setAsDictionary($template, $clone, $nullable);
 		return $this;
 	}
 	
@@ -1051,21 +1042,18 @@ trait Evaluators
 	 * @see \Dracodeum\Kit\Interfaces\Arrayable
 	 * @param \Dracodeum\Kit\Primitives\Vector|null $template [default = null]
 	 * <p>The template instance to clone from and evaluate into.</p>
-	 * @param bool|null $clone_recursive [default = null]
-	 * <p>Clone the given value recursively.<br>
-	 * If set to boolean <code>false</code> and an instance is given, then clone it into a new one with the same values 
-	 * and evaluator functions, but not recursively.<br>
-	 * If not set, then the given value is not cloned.</p>
+	 * @param bool $clone [default = false]
+	 * <p>If an instance is given, then clone it into a new one with the same values and evaluator functions.</p>
 	 * @param bool $nullable [default = false]
 	 * <p>Allow a value to evaluate as <code>null</code>.</p>
 	 * @return $this
 	 * <p>This instance, for chaining purposes.</p>
 	 */
 	final public function setAsVector(
-		?Vector $template = null, ?bool $clone_recursive = null, bool $nullable = false
+		?Vector $template = null, bool $clone = false, bool $nullable = false
 	): object
 	{
-		$this->getEvaluatorsManager()->setAsVector($template, $clone_recursive, $nullable);
+		$this->getEvaluatorsManager()->setAsVector($template, $clone, $nullable);
 		return $this;
 	}