+++ src/Dracodeum/Kit/Utilities/Text.php
@@ -1179,7 +1179,7 @@ final class Text extends Utility
 	final public static function parse(string $string, array $fields_patterns, $options = null): ?array
 	{
 		//initialize
-		$options = Options\Mparse::coerce(Options\Parse::coerce($options), false);
+		$options = Options\Mparse::coerce(Options\Parse::coerce($options), true);
 		$no_throw = $options->no_throw;
 		$options->no_throw = true;
 		$options->keep_nulls = true;